requirements.txt
mysql-connector
Flask==0.10.1
requests==2.12.1
watson_developer_cloud
